+++ packages/react/src/components/UIShell/components/SharkFinIcon.tsx
@@ -44,6 +44,7 @@ export const SharkFinIcon: React.FC<SharkFinIconProps> = ({
       width="4"
       height="4"
       viewBox="0 0 4 4"
+      aria-hidden="true"
       fill="none"
       xmlns="http://www.w3.org/2000/svg">
       <g clip-path="url(#clip0_519_52879)">