+++ packages/react/src/components/StylePicker/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.21.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.20.0...@carbon-labs/react-style-picker@0.21.0) (2026-05-12)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.20.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.19.0...@carbon-labs/react-style-picker@0.20.0) (2026-05-11)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.19.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.18.0...@carbon-labs/react-style-picker@0.19.0) (2026-05-08)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.18.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.17.0...@carbon-labs/react-style-picker@0.18.0) (2026-05-06)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.17.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.16.0...@carbon-labs/react-style-picker@0.17.0) (2026-05-06)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.16.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.15.0...@carbon-labs/react-style-picker@0.16.0) (2026-04-27)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.15.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.14.0...@carbon-labs/react-style-picker@0.15.0) (2026-04-14)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.14.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.12.0...@carbon-labs/react-style-picker@0.14.0) (2026-03-25)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.13.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.12.0...@carbon-labs/react-style-picker@0.13.0) (2026-03-24)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.12.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.11.0...@carbon-labs/react-style-picker@0.12.0) (2026-03-09)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
 # [0.11.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.10.0...@carbon-labs/react-style-picker@0.11.0) (2026-01-14)
 
 **Note:** Version bump only for package @carbon-labs/react-style-picker